celery_test/requirements.txt
celery
redis
asgiref==3.7.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.2.0
contourpy==1.1.0
crispy-bootstrap4==2022.1
cryptography==41.0.2
cycler==0.11.0
defusedxml==0.7.1
distlib==0.3.7
Django==4.2.3
django-allauth==0.54.0
django-crispy-forms==2.0
filelock==3.12.2
fonttools==4.41.1
google-api-core==2.11.1
google-api-python-client==2.96.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.1
httplib2==0.22.0
identify==2.5.26
idna==3.4
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
networkx==3.1
nodeenv==1.8.0
numpy==1.25.1
oauthlib==3.2.2
opencv-contrib-python
Pillow==10.0.0
platformdirs==3.9.1
pre-commit==3.3.3
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.0.9
python-dateutil==2.8.2
python3-openid==3.2.0
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
sympy==1.12
torch==2.0.1
torchvision==0.15.2
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
